cerbes/pyproject.toml
[tool.poetry]
name = "cerbes"
version = "0.1.0"
description = ""
authors = ["rarnal <arnal.romain@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyjwt = "^1.7.1"
fastapi = "^0.60.1"
uvicorn = "^0.11.8"
pydantic = "^1.6.1"
meltingpot = {path = "../meltingpot"}

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
ipython = "^7.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
